+++ grout/exceptions.py
@@ -17,3 +17,25 @@ class QueryParameterException(APIException):
 
         def __str__(self):
             return repr(self.detail)
+
+
+GEOMETRY_TYPE_ERROR = ("Incoming Record geometry '{incoming}' does not match the " +
+                       "expected geometry type '{expected}' of the RecordType {uuid}.")
+DATETIME_NOT_PERMITTED = ("This field is not permitted for Records of RecordType {uuid}. " +
+                          "Please remove this field, or update the RecordType to set " +
+                          "the `temporal` flag to True")
+DATETIME_REQUIRED = ("This field is required for all Records of RecordType {uuid}. " +
+                     "Please include this field, or update the RecordType to set " +
+                     "the `temporal` flag to False.")
+DATETIME_FORMAT_ERROR = ('ISO 8601 formatted with timezone information. Please check ' +
+                         'that the URL is properly encoded.')
+BASE_MIN_DATE_RANGE_ERROR = "Value must be the same or earlier than '{max}'."
+BASE_MAX_DATE_RANGE_ERROR = "Value must be the same or later than '{min}'."
+
+# Date range errors for the Record model fields (occurred_from and occurred_to).
+MIN_DATE_RANGE_ERROR = BASE_MIN_DATE_RANGE_ERROR.format(max='occurred_to')
+MAX_DATE_RANGE_ERROR = BASE_MAX_DATE_RANGE_ERROR.format(min='occurred_from')
+
+# Date range errors for the Record filters (occurred_min and occurred_max).
+MIN_DATE_RANGE_FILTER_ERROR = BASE_MIN_DATE_RANGE_ERROR.format(max='occurred_max')
+MAX_DATE_RANGE_FILTER_ERROR = BASE_MAX_DATE_RANGE_ERROR.format(min='occurred_min')