+++ tests/test_date_filter.py
@@ -1,8 +1,7 @@
-from dateutil.parser import parse
 import json
 
 import mock
-
+from dateutil.parser import parse
 from django.test import TestCase
 from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.auth.models import User
@@ -10,9 +9,12 @@ from django.http import HttpRequest
 
 from rest_framework.test import APIRequestFactory, force_authenticate
 
-from grout.models import Record, RecordSchema, RecordType
+from grout import models
 from grout.views import RecordViewSet
-from grout.filters import DateRangeFilterBackend
+from grout.exceptions import (QueryParameterException,
+                              DATETIME_FORMAT_ERROR,
+                              MIN_DATE_RANGE_FILTER_ERROR,
+                              MAX_DATE_RANGE_FILTER_ERROR)
 
 
 class DateFilterBackendTestCase(TestCase):
@@ -27,9 +29,8 @@ class DateFilterBackendTestCase(TestCase):
                                                  '123')
 
         self.factory = APIRequestFactory()
-        self.queryset = RecordSchema.objects.all()
 
-        self.item_type = RecordType.objects.create(label='item', plural_label='items')
+        self.item_type = models.RecordType.objects.create(label='item', plural_label='items')
 
         self.item_schema = {
             "$schema": "http://json-schema.org/draft-04/schema#",
@@ -49,50 +50,125 @@ class DateFilterBackendTestCase(TestCase):
             "required": ["id", "name"]
         }
 
-        self.schema = RecordSchema.objects.create(record_type=self.item_type,
+        self.schema = models.RecordSchema.objects.create(record_type=self.item_type,
                                                   version=1,
                                                   schema=self.item_schema)
 
-        self.a_date       = parse('2015-01-01T00:00:00+00:00')
-        self.a_later_date = parse('2015-02-22T00:00:00+00:00')
+        # Define five chronological points in time in order to test range filtering.
+        self.min_date = parse('2015-01-01T00:00:00+00:00')
+        self.lower_quartile_date = parse('2015-01-15T00:00:00+00:00')
+        self.mid_date = parse('2015-02-22T00:00:00+00:00')
+        self.upper_quartile_date = parse('2018-07-30T00:00:00+00:00')
+        self.max_date = parse('2018-08-01T00:00:00+00:00')
 
-        self.early_record = Record.objects.create(
-            occurred_from=self.a_date,  # A DATE
-            occurred_to=self.a_date,
+        # The first Record spans the first and second points in time.
+        self.min_to_mid_date_record = models.Record.objects.create(
+            occurred_from=self.min_date,
+            occurred_to=self.mid_date,
             geom='POINT (0 0)',
             schema=self.schema,
             data={}
         )
-        self.later_record = Record.objects.create(
-            occurred_from=self.a_later_date,  # A LATER DATE
-            occurred_to=self.a_later_date,
+        # The second Record spans the second and third points in time.
+        self.mid_to_max_date_record = models.Record.objects.create(
+            occurred_from=self.mid_date,
+            occurred_to=self.max_date,
             geom='POINT (0 0)',
             schema=self.schema,
             data={}
         )
+
+        # Create Records for a nontemporal RecordType.
+        self.nontemporal_record_type = models.RecordType.objects.create(
+            label='Nontemporal',
+            plural_label='Nontemporals',
+            temporal=False
+        )
+        self.nontemporal_schema = models.RecordSchema.objects.create(
+            record_type=self.nontemporal_record_type,
+            version=1,
+            schema={}
+        )
+        self.nontemporal_record = models.Record.objects.create(
+            schema=self.nontemporal_schema,
+            geom='POINT (0 0)',
+            data={}
+        )
+
         self.view = RecordViewSet.as_view({'get': 'list'})
 
     def test_valid_datefilter(self):
         """ Test filtering on dates """
-        self.assertEqual(len(self.queryset), 1)
+        # Define a set of filters along with the records that they should match.
+        tests = [
+            {
+                'filters': {'occurred_min': 'min_date'},
+                'matches': ['min_to_mid_date_record', 'mid_to_max_date_record']
+            },
+            {
+                'filters': {'occurred_min': 'upper_quartile_date'},
+                'matches': ['mid_to_max_date_record']
+            },
+            {
+                'filters': {'occurred_max': 'min_date'},
+                'matches': ['min_to_mid_date_record']
+            },
+            {
+                'filters': {'occurred_max': 'mid_date'},
+                'matches': ['min_to_mid_date_record', 'mid_to_max_date_record']
+            },
+            {
+                'filters': {
+                    'occurred_min': 'min_date',
+                    'occurred_max': 'lower_quartile_date'
+                },
+                'matches': ['min_to_mid_date_record']
+            },
+            {
+                'filters': {
+                    'occurred_min': 'upper_quartile_date',
+                    'occurred_max': 'max_date'
+                },
+                'matches': ['mid_to_max_date_record']
+            },
+            {
+                'filters': {
+                    'occurred_min': 'lower_quartile_date',
+                    'occurred_max': 'upper_quartile_date'
+                },
+                'matches': ['min_to_mid_date_record', 'mid_to_max_date_record']
+            },
+            {
+                'filters': {
+                    'occurred_min': 'min_date',
+                    'occurred_max': 'max_date'
+                },
+                'matches': ['min_to_mid_date_record', 'mid_to_max_date_record']
+            },
+        ]
 
-        req1 = self.factory.get('/foo/', {'occurred_max': self.a_date})
-        force_authenticate(req1, self.user)
-        res1 = self.view(req1).render()
-        self.assertEqual(json.loads(res1.content.decode('utf-8'))['count'], 1)
-
-        req2 = self.factory.get('/foo/', {'occurred_max': self.a_later_date})
-        force_authenticate(req2, self.user)
-        res2 = self.view(req2).render()
-        self.assertEqual(json.loads(res2.content.decode('utf-8'))['count'], 2)
+        # For each filters/matches pair, test to make sure that the date filters
+        # performs correctly.
+        for test in tests:
+            filters = {key: getattr(self, val) for key, val in test['filters'].items()}
+            req = self.factory.get('/foo/', filters)
+            expected_count = len(test['matches'])
+            force_authenticate(req, self.user)
+            res = self.view(req).render()
+            self.assertEqual(json.loads(res.content.decode('utf-8')).get('count'),
+                             expected_count,
+                             test)
 
     def test_missing_min_max(self):
-        """Test that forgetting both `occurred_min` and `occurred_max` returns all records."""
+        """
+        Test that forgetting both `occurred_min` and `occurred_max` returns all records,
+        including the nontemporal record.
+        """
         missing_range_req = self.factory.get('/foo/')
         force_authenticate(missing_range_req, self.user)
         missing_range_res = self.view(missing_range_req).render()
 
-        self.assertEqual(json.loads(missing_range_res.content.decode('utf-8'))['count'], 2)
+        self.assertEqual(json.loads(missing_range_res.content.decode('utf-8'))['count'], 3)
 
     def test_missing_timezone(self):
         """Test that forgetting timezone information raises an error."""
@@ -101,8 +177,10 @@ class DateFilterBackendTestCase(TestCase):
         force_authenticate(missing_tz_req, self.user)
         missing_tz_res = self.view(missing_tz_req).render()
 
-        msg = 'Invalid value for parameter datetimes, value must be ' + DateRangeFilterBackend.ERR_MSG
-        self.assertEqual(str(json.loads(missing_tz_res.content.decode('utf-8'))['detail']), msg)
+        # Both values have to be strings in order for the exception to fit
+        # the equality test.
+        self.assertEqual(str(json.loads(missing_tz_res.content.decode('utf-8'))['detail']),
+                         str(QueryParameterException('occurred_max', DATETIME_FORMAT_ERROR)))
 
     def test_range_parse_errors(self):
         """Test that parse errors get thrown when the date range is improperly formatted."""
@@ -111,13 +189,32 @@ class DateFilterBackendTestCase(TestCase):
         force_authenticate(bad_min_req, self.user)
         bad_min_res = self.view(bad_min_req).render()
 
-        msg = 'Invalid value for parameter occurred_min, value must be ' + DateRangeFilterBackend.ERR_MSG
-        self.assertEqual(str(json.loads(bad_min_res.content.decode('utf-8'))['detail']), msg)
+        self.assertEqual(str(json.loads(bad_min_res.content.decode('utf-8'))['detail']),
+                         str(QueryParameterException('occurred_min', DATETIME_FORMAT_ERROR)))
 
         # Test a bad `occurred_max` parameter.
         bad_max_req = self.factory.get('/foo/', {'occurred_max': 'foobarbaz'})
         force_authenticate(bad_max_req, self.user)
         bad_max_res = self.view(bad_max_req).render()
 
-        msg = 'Invalid value for parameter occurred_max, value must be ' + DateRangeFilterBackend.ERR_MSG
-        self.assertEqual(str(json.loads(bad_max_res.content.decode('utf-8'))['detail']), msg)
+        self.assertEqual(str(json.loads(bad_max_res.content.decode('utf-8'))['detail']),
+                         str(QueryParameterException('occurred_max', DATETIME_FORMAT_ERROR)))
+
+    def test_invalid_range(self):
+        """
+        Test that min > max raises an error when both ends of the date range
+        filter are provided.
+        """
+        min_gt_max = {
+            'occurred_min': self.max_date,
+            'occurred_max': self.min_date
+        }
+        min_gt_max_req = self.factory.get('/foo/', min_gt_max)
+        force_authenticate(min_gt_max_req, self.user)
+        min_gt_max_res = self.view(min_gt_max_req).render()
+        self.assertEqual(str(json.loads(min_gt_max_res.content.decode('utf-8')).get('occurred_min')),
+                         MIN_DATE_RANGE_FILTER_ERROR,
+                         min_gt_max_res.content)
+        self.assertEqual(str(json.loads(min_gt_max_res.content.decode('utf-8')).get('occurred_max')),
+                         MAX_DATE_RANGE_FILTER_ERROR,
+                         min_gt_max_res.content)